Moved Appdatabase.kt and finished the part of posting items in navhost
