+++ app/src/main/java/com/example/monashMP/data/database/AppDatabase.kt
@@ -1,4 +1,4 @@
-package com.example.monashMP.data
+package com.example.monashMP.data.database
 
 import android.content.Context
 import androidx.room.Database
